stop overwriting the "focus" selection
now it determines the focus options before generating the map, and determines at that time whether it shuold keep the old selection or revert to the default agen.  it then passes both the list of options and the current selection back to the gui thread so that they can be updated in a way that's always consistent with the map.  this fixes the previus bug wherein the focus option would keep getting reset to "Continent 1".