adjust map cropping
whole world maps don't actually need to show the whole planet if there's a big ocean on one of the sides.  and labels shouldn't try _too_ hard to avoid lakes.  ideally it would ignore the lake when doing the strait skeleton but then avoid them when doing the actual label placement, but that's way too complicated for now.  I also adjusted the scale of greebling and the graticule.