change a confusing variable name
I think callnig this class Name instead of Word makes some sense because it makes it clear that it gets adapted and used in different languages, but it's really confusing for Lect.getName() to not be the name of the Lect, like Civ.getName() is.  so Word is better, and now I have space to add an actual Lect.getName() function.