add missing license text
the MIT license requires that the license text accompany any copy of the software, but the copy of plotly that I downloaded didn't have it.  they reserve that right, but I don't, so I added the license text just to be safe.  I also deleted the line that sed "All rights reserved" since, uh, they aren't?  like the line right under that is a denial of any such reservation of most of their copy rights.