sprinkle some little pictures on the map
wow, this one was way harder than I expected.  I mean, mostly around establishing a robust method for loading assets that I want to use in the map.  unlike TS files, SVG files can't be imported and must be obtained using fetch, which is a huge pain on the worker thread because it's asynchronous.  I can't build the map and then drop in the resource when it's ready like I did originally for the windrose; once the map is sent it's sent.  so I can't start generating the map until the resources are loaded.  but there's no way in Javascript to block until fetch is done; it _always_ queues fetch() up to execute _after_ the current process is over, meaning if you block that thread it will never get done.  this design choice baffles me; why is there no option to just do it on the curent thread‽  anyway, I apparently need to have a variable that tracks whether the worker is ready, and it needs to defer any messages it receives while fetch is going, and then when it finishes and the worker is ready it needs to go back and respond to all of those deferred messages.  it's _really_ annoying, especially since the Worker class clearly already has some kind of message queueing bilt-in, so why can't there just be a method that says "I'm not taking messages now; hold onto them as if I were in the middle of another callback"?  whatever, it works now.  I mean the actual texture thing is minimal, but it onestly looks better than I expected.