simplify poisson disc sampling algorithm
I got rid of the grid cell queue thing because I did some monte carlo calculations and found that it wasn't actually saving me much time (and was also adding considerable overhead).  now, I just draw each sample from the bounding box of the domain.  the especially nice thing about this is that it no longer has a propensity to put a lot of trees right on the borders of the region, and especially on convex corners, which makes it look just a little bit nicer.  I'm also no longer aiming for a maximal distribution of points, because I realized it's kind of inefficient to try to do that with my current tecnikes, and also I think it's kind of nice to have some randomness in there.