basic coastal shading
I want to have the thing where the region of ocean near land is horizontally hatched.  I have implemented the basics needed to accomplish that here.  mostly, I wrote a janky curve offset algorithm.  it doesn't clean its result up like I would like it to, and I also haven't written the actual hatching part, but that can come next.

I made a suprising number of other changes in support of this offset function, including a decimation function, a line-segment-to-point distance function, a reversePath function, and a utility function in Chart that calls projectPath, convertToGreebledPath, and outline together with the correct default arguments.  I also added a bunch of tests, including a test for polygonize().