make offset() work with holes
it turns out there was a way simpler way to do this, which came to me today as I was brushing my teeth.  I actually didn't need to make the last commit, but I was so close to implementing the whole "removing holes" thing that I figured it would be nice to just have it in the record.  and I'm keeping some of the refactoring I did there.  but it turns out this actually works great with holes if I just reverse the polarity of concave corner arcs.  from a visual geometric standpoint, it felt more logical to have them backtrack, as if there's a tiny fillet inside every corner, but interestingly it's actually somewhat more mathematicly elegant if it keeps going and does a big loop to get to its new location.  and that makes holes work.  :shrug: