hide texture elements outside of the map area
now the texture elements no longer extend outside the border of the map.  I accomplishd this by adding a white border to the region that's part of the SVG but outside the map region (this will also help it convert to PNG better.  someday I may add decorative detail to that.), and by adjusting the CSS on the map itself so that teh SVG only fills space equal t oits viewbox rather than expanding to fill the whole available space.