real pictures
there are a few adjustments that I need to make, to be sure.  the pine trees are as tall as mountains – they should be made smaller, and also pointier (as they currently look like nondescript tall blobs).  the monkeypod trees could stand to be taller, and the banyan trees are hard to parse; their trunks should be farther apart.  the shrubs look kind of like cracks, so I'd like to round out their edges a bit.  in general, I think I could do with more irregular spacing, especially when it comes to grass and shrubs.  and, of course, I still need to adjust the spacing so things don't overlap with coastlines.