make texture avoid edges
now poissonDiscSample() places samples such that they are never within r/2 of the edge.  all I had to do was use offset() and contains() to erode the feasible region by that amount.  this cleanly takes care of the issue I had noticed before where samples tend to cluster right on the boundary, also makes it so that mountains no longer intersect with coastlines.  now I just need them to also not intersect with borders and rivers, which should be easy with this new framework.

the one issue is that sometimes it doesn't place as many symbols as I would like when the minimum spacing is largeish.  I think it would look better if I allowed symbols to have varying radii.

Note that it also might be faster if I decide to be smarter about the feasibility grid.  because of the way I do offsets, this is going to mark a bunch of cells outside of the true polygon as "borderline" outside of the offset polygon.  I could excise those, which might alleviate the need for a substantial number of calls to contains().  but let's push this for now and see what the actual slowdowns are.