fix arbitrarily complex lines near pole
there was a rare bug I found where the path projection algorithm would fail when a coast passed too close to the outer pole in a conic projection, because it could get inflated to arbitrary scale which could require arbitrarily many points to sample sufficiently.  so now it doesn't try to supersample lines at all if they're outside of the map area.