make textures denser
what in theory should have been possible by simply turning up some numbers in texture_mixes.ts ended up having quite a few knock-on effects.  I had to adjust outline to take a set instead of any iterable so that it made sense to convet to set in projectedOutline, which made it possible to check for region area zeroness in projectedOutline, which I wanted to do so I would no longer have to check that in generateTexture because I was refactoring generateTexture to streamline the inside of the nested for loops because now I treat mountains as different from hills (since all mountains have the same texture, you can get much better results by lumping them all together into a single poisson disc region).  this means that the coloration needs to be handled separately since the mountains will have different colors on a physical map, so now colorSchemeName doesn't need to get passed to generateTexture either.  also we don't need to do lakes separately when we're doing the physical coloration and we haven't for a while because sea color is no longer an independent parameter.

overall I think it looks much better, tho the mountains are still not where I want them to be.  I think I need to mix some smaller mountains in.  but that may be a post-1.0 thing.