make shaded relief less dark
I change it to use nothing for flat ground, transluscent black for downward slopes, and transluscent white for upward slopes, rather than before where it used transluscent black for everything except extremely upward slopes.  naturally it doesn't show upward slopes as well now, especially for white backgrounds, and is somewhat harder to read overall.  eventually I'll add more vertices to the triangulation, which I think will help, but for now it's fine; at least you can actually see the flat areas.  also, I think it looks better in general for physical maps.

I also added a clip path so that it doesn't extend into the ocean or lakes anymore.  onestly I'm pretty annoyed I had to do that since I haven't used any clip paths yet, and the only reason I can't simply put the ocean on top of the shading is that the trees need to stick out over the ocean sometimes and the shading needs to be on top of the trees (but how often are people actually going to use both shading and trees?).  idk, I reserve the right to reorder things and just not worry about having trees on top of the shading if the awkward half-shading of coastlines gets to me, but for now it's fine.  also worth noting that this will be all be less of an issue when I add more vertices.

I also improved the scaling overall so it doesn't try to make everything dark to compensate for one little vertical triangle.