make map rectangularity depend on projection
I recognize I'm waffling on this, but on further consideration this really isn't a useful option for the user.  why would anyone ever want a rectangular map with the Bonne projection?  it just looks ugly when you zoom too far out!  my goal is not to give the user complete control; it's to let them choose their preferences and generate a result that looks nice.