move shaded relief below ocean
yeah the way I did it was silly.  this new way looks worse for with texture and relief, but that was going to look bad anyway!  what's important is that it looks better with just relief, and it means I don't have to use any clip paths, which don't always import nicely and make the SVG more complicated.