fix a contains() edge case
man, I thaut I was done with these!  just goes to show, you never run out of edge cases.  this one was a roundoff issue with arcs whose endpoint was at the same y-value as the point being tested.  I also noticed a potential future issue to do with such arcs that also pass thru the same horizontal line at a different point.  also, this didn't need to be exported.