fix more roundoff issues
this one was because I recently added a check to throw an error when points being tested by contains() are out of the bounds of the Domain object, which does happen because of this weird thing I do to try to prevent roundoff issues inside the Domain object.  the error didn't really need to be thrown, but I figure the safest solution is to simply set these points to match the new slitely modified domain bounds.