refactor chart.ts
I removed the Chart object and combined the constructor and depict() into a single function that takes a world and spits out an SVG.  I then tried to push as much of the content of that function into subfunctions as I could.  this makes it a little easier to read, I think, but more importantly it will be very useful for profiling, which I do intend to do at some point.

I had to get rid of the COLOR_BY_TECHNOLOGY debug option, for now at least, because it didn't cleanly fit into my new framework, and I didn't want to write too much code for a feature that the user can't even access.  I also had to delete chart.test.ts because it was not well equipped for this new scheme and I don't think it's terribly necessary.

one kind of weird necessary modification that I think is worth noting: I changed the initial country index from 0 to 1 so that I can pass the list of country colors to a function with the un-shaded white also part of it by simply declaring that 0 means none.