fix some more plotting errors
varius minor fixen.  when you transformInput in transformInput you should use the projection's domain, not its φMin and λMin (these may differ for the Bonne projection; it's not a big deal but breaks that new domain check I added to contains()).  when it trips that check for too many intermediate points, it now errors in a way that still returns an actual map.  there was this assertion in intersection() that went like "if (encompasses(edges, segment) != encompasses(edges, segment))" (i.e. both sides were the same) and it's been like this for years so I didn't look back far enuff to see when it got like that or what it's even supposed to be (presumably one of them is supposed to be the _next_ segment rather than the last segment) so I just deleted it.  I eased this roundoff check thing in localizeInRange because it was actually _causing_ roundoff issues in this one niche case with the gradient.  better to leave numbers where they are unless there's a good reason to snap them to the bounds.