make contains() more robust with arcs
I had learned this lesson years ago with getMapCrossings(): while it is possible to check if a point is on an arc with just a single comparison in theory, that method becomes highly susceptible to roundoff errors when the arc is small.  I forgot about that when I spun out getAllHorizontalLineCrossings(), and fell prey to the same plotting error I first fixed all those years ago.