better centering on conic projections
I realized that it looked really ugly to have conic projections where the central meridian was not centered on the map.  the problem was that if you want to center it, you need to either accept that the landmass being shown won't necessarily be centered, or solve the question of how to choose the central meridian so that you can center it and the landmass simultaneusly.  this commit solves that problem.  it looks great!  except it doesn't always work.  but it usually works, so I'm pushing it, and I'll push a fix for those edge cases later.