fix triangulation edge case
this would occasionally come up for toroids, where a point is right on the edge between two triangles, so it tests positive for containment by one of them but then tests negative for containment by all of its children, resulting in the error message "no eureka tingo da indu".  I realized that I don't actually need to check edges multiple times; when I'm drilling down to see which of a triangle's children owns a point, I can just check the edges _between siblings_, which prevents that from possibly being an issue (and it's fine either way for the triangulation because if a point is that close to an edge than the edge is basicly garanteed to flip.

main actual code changes are keeping track of which edges are between siblings.  while I was in there I also did some basic cleanup, like rephrasing things to be in a more intuitive order.  I also added an "enabled" boolean to triangles to use rather than just knowing that a triangle is active if its children are null and inactive if its children are non-null (there were a bunch of statements that set children to [] to disable a triangle that were horribly unintuitive).