make straight skeleton work for concaves
when the surface mesh is too coarse, polygons may come out slitely concave.  when this hapenned it caused the straight skeleton algorithm to fail.  now, tho, it will still work in that scenario.  it won't produce a medial axis the way it does for convex polygons, so it's not quite as optimal for greebling purposes, but it seems that most polygons come out convex regardless so I'm not too worried.