include ice in "whole world" maps
I made it stop excluding ice from maps of the whole world.  as practical as it is to leave it out, given how common it is to leave Antarctica out of real-world maps, I think users will be confused to choose "whole world" and then see that there's clearly still land going off the page.  especially if a country has claimed any of that land.  and especially especially on toroidal planets where they may be confused about the extent of the world in general.  and anyway, it's not like it's rare to see antarctica on a world map – it's still probably on the majority of them.