improve continent definition
I tweaked the continent finding algorithm a bit.  in general, the continents are now allowed to be smaller but are merged more aggressively.  this helps with some big islands I had that seemed like they should be a continent but weren't part of any continent, and with seemingly arbitrary divides that ran right thru the middles of continents.

in addition, there is now some amount of self-adjustment when the initial algorithm produces more than 8 (which I've never seen tbf) or fewer than 2 (which was all too common).  continents are defined by the borders between them, so it doesn't make logical sense to have just one continent, and the code now reflects that.