fix some map centering bugs
it wasn't accounting for wrapping of 'Φ' segments correctly, which made it place inland seas in the center of the map rather than on the sides.  and it wasn't excluding ocean when it did the weited average method.  and it wasn't rotating the tile representation of the land properly when doing the central parallel, so toroids would be doubly rong when they had to fall back on the weited average method.