disconnected colonies gain independence
now, when a colony becomes disconnected from its capital due to war in the homeland, rather than everyone in the colony spontaneusly dying (as happened previusly), it becomes an independent nation.  this required the creation of a new helper function for spawning new civs.  I also changed the way Civ.border works; I'm not sure if it was necessary for this colony thing but I wasn't convinced that what it was doing before was correct, so I made it be just a set of tiles on the border without a mapping to sets of neibors that are outside the border.  much easier to maintain this way.  I also ended up having to make the random number generator a public instance variable of World to help with spawning the new guys, which I think is reasonable given that a single random number generator was being passed around for all the history simulation stuff anyway.