fix bug with predecessor clarification
the name of the country should always come first.  the history message for a new civ spawning in the wild didn't follow this convention, and the "which is the predecessor of" message was being inserted into the wrong part of the sentence as a result.  as a further gard agenst this kind of thing, I also now check whether the historical event participant is a country or a people so that it won't look quite so bad even if I make the same mistake agen elsewhere.