give empires a firmer deathtimer
I didn't like the possibility of an empire spawning with a really high military strength and just sticking around indefinitely.  a decay time of 1000 years was way too slow, and in my tests it seemed like this was far too common.  the issue is that if an empire is too strong then any rebellions will automaticly fail, and then they just spread to cover everything.  this way, it decays linearly instead of exponentially, so after 1000 years the effect is similar, but after 2000 years it's basicly garanteed to go away.