ensure Civ IDs are unique
this was a pretty rare case, but now that I've implemented independence as a result of territorial disruption, it is possible for spawning a Civ to spawn more Civs, and when that happens, it messes up the IDs and makes them non-unique.  changing the order of these calls, tho, so that it increments the ID _before_ passing it to the constructor, fixes that rite up.