only process rebellions that stand a chance
it used to be that many rebellions would happen but only around half of them would end up holding their territory for more than 0 seconds.  now, whenever it starts a rebellion, it checks if the breakaway country has enuff strength to beat its predecessor.  if so, the rebellion proceeds as normal.  if not, the rebellion is cancelled and the uprising is quashed.

to do this, I ended up moving a bunch of stuff from the Civ constructor to capital.  so now the capital is instantiated as null, and it only sets the capital and generates a new culture and records the event in the annals of history if the new stat's strength is high enuff.  this is kind of awkward; it means that capital can't be readonly anymore even tho practicly speaking it's still getting set once rite after the Civ is instantiated, and it means that I have to be careful not to call, e.g., getName() on a Civ until it conquers its capital.  but this was the best way of which I could think to make it possible to check the validity of the new Civ halfway thru and cancel its constructor if it's not valid.  besides exceptions, anyway, which I did consider.  but exceptions aren't great in JS, and it also was going to be tricky with states that breakaway as a result of territory loss (because those actually need to get created regardless of how their strength compares to their parent).