add an age to the tech tree
it felt weird to call civilizations with catapults and steel weaponly "iron age", and to have it then skip strait to being "medieval".  so now there's a "classical" tech level in the middle.  I also changed the Japanese names of the eras to Japanese era names, which I think will be more understandable to a Japanese audience.  it's a little weerd since I don't have all of the era names; I suppose I could add more granularity and have it not appear in the English and Spanish translations but give different Japanese era names in Japanese, but I don't really want to do that.  after all, was the technology in the late Heian, Kamakura, Muromachi, and early sengoku periods really different enuff that one would be able to tell the difference by tech alone?  they all had stirrups but not muskets.

I also reframed stirrups as "fighting with horse-mounted warriors" and not "getting around on horseback"; it's not like they invented horseback riding in 300 CE.