make cultures independent of civs
now a single culture object manages all of the speakers of a given language even if they span multiple countries.  this fixes some issues I was having with seemingly duplicate cultures.  it's worth noting that I had to get rid of the feature where minorities in countries ruled by sedentary people aren't allowed to be nomadic, but whatever; I don't think that was very necessary, and it wasn't very realistic either.