fix language name bugs
I forgot to remove diversity when I stopped defining it (thanks for the warning, JS) and I also messed up the way it sets the name of languages so that they never matched the people names.  some of the language names are kind of long now, so I have to deal with that somehow.