new transcription style
I realized that while the anglicized spelling system is neat and common in north american languages, english speakers will probably be most familiar with something that uses the standard five-vowel system and some digraphs for common nonenglish sounds, but very few diacritics.  so I reworkd the latin transcriptions such that they don't all have to be tied to a specific language and this new one is called just "simple".