no uvular laterals
it's a little bit shocking to me that there was an illegal sound still hiding in my process list after all this time.  I suppose I've been reducing the number of processes, which probably makes this kind of thing more likely to show up in tests which is probably good and probably increases the probability that there aren't more of these.