fix words not capitalizing correctly
I accidentally made it so most languages capitalize words but not phrases.  obviusly unintentional.  not that the behavior is better now, mind you – it happened to look better when it only capitalized phrases because names are phrases while common nouns are words.  but that's just a coincidence, and not one I want to be forced to rely on.