languages take time to diverge
now there's a difference between when two _cultures_ are considerd different (which takes 200 years) and when two _languages_ are considerd different (which now takes 500 years0.  and if two cultures speak the same language (which is now therefore possible) then it says so in the factbook because it uses the languages' name from 500 years ago, not the name inferred from the state now.  I think this is working correctly.  actually I'm not positive because whenever I change DEVIATION_TIME it throws off the random seed, which is not supposed to happen, so I have to look into that, but I'm committing this now because it's probably not related to this change.