make mountains higher
I made a bunch of tweaks to the terrain generation, focusing on altitude generation.  mountains are higher in general, but they're also more consistent in their height, so you don't get too many extremely high mountains anymore.  I increased the amount of variation on continental interiors to be more realistic, and scaled the lake threshold and river canyon threshold to match.  I also turned down ocean depth variation because I think the sea floor is generally more level than continental interiors.    I redefined altitude initially so that average land height is 0 and average ocean depth is -4km (this ceases to matter after the ocean is added so it doesn't change any behavior but I like it better this way).  I simplified the mid-ocean rift formula because there was a lot of complexity to account for a rare edge case that I don't think actually mattered at all.