scale a couple other things for altitude
now that mountains are on average twice as high, we need the temperature to vary haff as much with altitude and we need the orographic effect to apply higher.  but the orographic doesn't have to quite scale up proportionally because that altitude is mostly based on first principles and I've always wanted more orography.  since I didn't scale it up all the way (and because I'm about to add more mountains but don't tell anyone I sed that) moisture is a little harder to come by, so the evaporation thresholds needed to change.