simplify river flow rate criterion
I made it so that rivers can no longer spawn in deserts by removing this free +1 water generation rate that I had initially added.  I have mixed feelings about it, but I do prefer it without that factor.  now there's more swing in how many rivers are visible on any map based on which biome it's in, but most importantly there are no longer lakes spawning in the middles of deserts unless there's a river transiting the desert to supply it.  it's more realistic and I was sick of seeing those improbable lakes.  I think the solution to the issue of too many rivers is to have something intelligently tune the threshold based on how many are visible, to show more of a river's full length even if the tip is low flow rate, and to smooth them.  but that's all post-1.0 work.