maritime empires
now, when a country is born on the coast, it has a small chance to become _thalassocratic_, which gives it a 10× boost to all ocean combat and a 3× boost to all coastal combat.  I'm hoping to get more countries that span multiple landmasses, and that stretch along long coastlines this way.  so far it seems to be working decently well.