brighten political map colors
I increased the saturation, changed from CIECAM02-UCS to Oklab in the hopes of reducing the amount of blue, and then also removed all blues that were too close to the water color.  I also lightened the water to further reduce the confusion between blue countries and ocean.  this does unfortunately mean I have fewer colors to work with in total, but it's probably for the best.