consolidate calls to rasterInclusion()
it turns out this function is the slowest part of the map generation right now, and I'm probably calling it dozens more times than I need to.  it should be faster to consolidate all the walls I have to avoid in poissonDiscSample and just call rasterInclusion() once on the whole thing.