revamp containsRaster()
I totally rewrote rasterInclusion() and renamed it to containsRaster() to reflect the large amount of change.  now, there's no more flood-filling.  instead, it automaticly sets spaces to IN when it finds the horizontal comb crossings.  this allows me to stack multiple inclusion rasters on the same data structure more easily, allowing me to consolidate cellFeasibility and cellForbidenness.  more importantly, it allows me to add more information to the return value.  now, when a cell is borderline, it tells you _which parts_ of the feasibility polygon are touching it.  this makes it more efficient to check candidates because you no longer have to check the whole entire 10000-segment polygon.