make texture denser
it kind of seems like the last time I shrunk the texture I didn't shrink the diameters in concert?  so if everything seemed annoyingly sparse, that was one reason why.  anyway, I made all of the textures smaller and re-set their diameters, so it looks much fuller now than it did.  the mountains still aren't where I'd like them to be, but I think they're 1.0-worthy.