make jungle more traversable
now it's still less fertile to civilization than plains or forest but not actively repulsive to it.  like yes the amazon and central africa are some of the most isolated parts of the world but also like india and indonesia have plenty of civilization.  idk maybe indonesia is just from the thalassocracies.  but more importantly, I just prefer having the map filled out even if I think it's not  necessarily as realistic.