adjust color scheme again
I really need to just write a smarter color placement algorithm because I have spent way too much time trying to design a colormap to work with my dumb scheme.  whatever, I think this looks nicer because now the text is easier to read because the trees are lighter because the colors are lighter because the ocean is lighter.