fix some bugs related to cataclysms
apparently I don't test this feature enuff.  it was trying to erase humanity from tiles that were already uninhabited like deserts and oceans, which was a probem, and also when peoples' homeland got destroyd in a cataclysm then they were confused as to whether they counted as expats or not.