make lines behave better near poles
I discovered a new bug, where ungreebled segments near the pole were so curved by the coordinate system that they could flip the sign of a tile, or cause it to self-intersect.  this only seemed to be noticeable when COLOR_BY_TILE was turned on, but it's entirely possible it would have manifested in regular maps eventually.