update test for recent bugfix
sorry.  Maybe it was bad that I wrote a test that takes 60 seconds to run because there's no way I'm going to run the full test suite every time I make a commit now.  fun fact, Jest froze when it hit this test failure; it's been a really long time since it's done that!