improve Anglicization
I only just noticed this but it turns out english doesn't really use e + consonant + e at the end of a word to spell long e except when the consonant is "m", "n", or "v".  like, there are greco-latin words like "allele" and obscure words like "fete", but in general words like "fele" and "affete" look unnatural and don't feel like they should be pronounced with a long e.  instead it really aut to be "feel" and "afeet".  it's bizzare to me because all the other vowels can do this as I was told in elementary school, but e specificly just doesn't.  (credit to Coleman Carl Ellis for the "allele" example)