change html slightly
on mobile Firefox the interword spacing is bonkers and I can't figure out why.  it's fine on desktop with Firefox and it's fine with Chrome on mobile, but Firefox on mobile is a problem on multiple devices.  I'm hoping one of these things will have an effect even tho it's not clear to me how either of them would.