make fonts more consistent
it turns out that this fancy emoji font I've been trying to use for my emoji (Noto Color Emoji) is more commonly supported than Noto Sans, but also defines an emoji-wide space.  meaning that on devices without Noto Sans installed it falls back on Noto Color Emoji for the _spaces_, which makes the spaces all huge.  unfortunately it seems there's no way to use it without this happening besides specifying that the emoji specifically use Noto Color Emoji and everything else including spaces uses regular fonts, which doesn't help me because the only emoji I use are in the language dropdown where that kind of fancy formatting is not permitted.  instead I'm just going to remove it and assume users' browsers' default font has flag emoji.  I also changed the info icon to use an SVG instead of the "ⓘ" character, because the font I was using to render that one (Segoe UI Symbol) is not present on all devices and it seems the typical default fonts can't handle it.  it has an aria label so I think using an image for the actual "i" is fine.